+++ tests/asyncio/test_server.py
@@ -1,10 +1,12 @@
 import asyncio
 import dataclasses
+import gc
 import hmac
 import http
 import logging
 import socket
 import unittest
+import weakref
 
 from websockets.asyncio.client import connect, unix_connect
 from websockets.asyncio.server import *
@@ -22,6 +24,7 @@ from ..utils import (
     MS,
     SERVER_CONTEXT,
     LoggingTestCase,
+    skip_unless_reference_counting_collects,
     temp_unix_socket_path,
 )
 from .server import (
@@ -61,6 +64,27 @@ class ServerTests(EvalShellMixin, LoggingTestCase, unittest.IsolatedAsyncioTestC
                     str(raised.exception),
                     "received 1011 (internal error); then sent 1011 (internal error)",
                 )
+
+    @skip_unless_reference_counting_collects
+    async def test_garbage_collection_after_close(self):
+        """Closed connection is freed by reference counting in a server context."""
+        async with serve(*args) as server:
+            async with connect(get_uri(server)) as client:
+                await self.assertEval(client, "ws.protocol.state.name", "OPEN")
+
+                [handler_task] = server.handler_tasks
+                [connection] = server.all_connections
+                connection_ref = weakref.ref(connection)
+                gc.disable()
+                self.addCleanup(gc.enable)
+                del connection
+
+            # Closing the client causes the connection handler to return,
+            # which closes the connection on the server side.
+            await handler_task
+            del handler_task
+
+            self.assertIsNone(connection_ref())
 
     async def test_existing_socket(self):
         """Server receives connection using a pre-existing socket."""